[refactor] api, chat 모듈에서 동시에 사용되는 dto의 위치를 core 내부로 이동 (DirectMessageDto, UserSummaryDto) - ConversationQueryResult 위치 변경 (.../dto -> .../dto/response)
