+++ mopl-core/src/main/java/com/mopl/domain/user/dto/response/UserSummaryDto.java
